archive/legacy-2026-07-06-fresh-pull/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atom"
version = "0.1.0"
description = "AI-powered boxing drill coach"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "elevenlabs>=1.0.0",
    "pydub>=0.25.1",
    "edge-tts>=6.1.0",
    "mutagen>=1.47.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
atom = "atom.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/atom"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
